提交

+++ src/fps_overlay.h
@@ -69,7 +69,7 @@ private:
 
 	std::string text;
 
-	int last_speed_mod = 1;
+	float last_speed_mod = 1.0f;
 	bool speedup_dirty = true;
 	bool fps_dirty = true;
 	bool draw_fps = true;
